Feat/리스트페이지 디자인 구현 (#14)
* 🎨  style : tailwind.config.js , index.scss 파일에 커스텀 스타일 추가

* 🎨  style : 리스트페이지 반응형 구현

* 🚚  chore : CardList 컴포넌트 파일 경로 이동 및 파일명 변경
