+++ src/components/CardList/index.jsx
@@ -1,0 +1,40 @@
+import { useState, useEffect, useCallback } from 'react';
+import { getSubject } from 'api/subjects';
+import QuestionImage from 'assets/images/icons/ic_Messages.svg';
+
+const CardList = () => {
+  const [data, setData] = useState([]); // API 응답 데이터 저장
+
+  // API에서 데이터 받아오기
+  const userList = useCallback(async () => {
+    setData(await getSubject());
+  }, []);
+
+  useEffect(() => {
+    userList();
+  }, [userList]);
+
+  return (
+    <div>
+      <div className='grid gap-4 grid-cols-mobaileLow justify-center mx-6 mb-[31px] max-lg:hide-after-6 md:grid-cols-tabletLow md:mx-[130px] md:mb-[61px] md:gap-5 lg:grid-cols-pcLow lg:mx-[130px] lg:mb-[46px] pc:mb-10 pc:mx-[130px]'>
+        {Array.isArray(data.results) &&
+          data.results.map((item) => (
+            <div className=' max-w-[220px] border rounded-2xl border-gray-40 bg-gray-10 px-4 py-4' key={item.id}>
+              <div className='flex flex-col'>
+                <img className='rounded-full object-cover size-12 leading-6 mb-3' src={item.imageSource} alt='유저프로필' />
+                <div className='text-[18px] text-gray-60 font-normal mb-[30px]'>{item.name}</div>
+              </div>
+              <div className='flex items-center text-gray-40 text-[14px] leading-[18px] font-normal gap-1'>
+                <img className='size-4' src={QuestionImage} alt='질문아이콘' />
+                <div className='flex-1'>받은 질문</div>
+                <div>{item.questionCount}개</div>
+              </div>
+            </div>
+          ))}
+      </div>
+      <div className=''>pagenations</div>
+    </div>
+  );
+};
+
+export default CardList;
